fix: Resolve MCU command parsing and frontend button state issues
- Fix issuedAt parsing: Change to unsigned long long to handle large timestamps
- Fix BAD_PAYLOAD ACK: Extract requestId from invalid JSON for proper error reporting
- Fix frontend button state: Ensure opacity resets correctly after status changes
- Add ArduinoJson include for requestId extraction in error cases

Fixes:
- Protocol parser now correctly handles large issuedAt timestamps
- Backend can now match error ACKs with requestId for proper error handling
- Frontend button no longer remains inactive after completion
- False success toasts eliminated when MCU returns error ACKs

+++ firmware/src/protocol/Protocol.h
@@ -10,7 +10,7 @@ struct CommandResult {
     char requestId[37];      // UUID length (36 + null terminator)
     char command[16];
     char userId[64];
-    unsigned long issuedAt;
+    unsigned long long issuedAt;  // Use long long for large timestamps (milliseconds)
     bool valid;
 
     CommandResult() : valid(false), issuedAt(0) {
